backend/requirements.txt
django>=4.2,<5.0
djangorestframework
drf-spectacular==0.27.1
opencv-python-headless==4.8.1.78
pdf2image
psycopg2-binary
celery
redis
django-redis
django-cors-headers
python-dotenv
gunicorn
PyMuPDF==1.23.26
dj-database-url
django-ratelimit==4.1.0
python-magic==0.4.27
django-csp==3.8
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest~=8.0
pytest-django~=4.8
pytest-cov~=4.1
pytest-xdist~=3.5
numpy<2.0.0
tenacity<9.0.0
pytesseract

# Multi-layer OCR (PRD-19 robustification)
easyocr>=1.7.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
pytest-xdist>=3.5.0  # Parallel test execution
